Dissolve test_usecases.py
